+++ js/draw-thumbnails.js
@@ -1,5 +1,6 @@
 import {createPhotoDescriptions} from './mock/demo-data.js';
 import {drawBigPicture} from './draw-big-picture.js';
+import {modalOpen} from './modal.js';
 
 /**
  *
@@ -8,6 +9,7 @@ import {drawBigPicture} from './draw-big-picture.js';
  */
 const renderThumbnails = function (quantity) {
   const thumbnailsBlock = document.querySelector('.pictures');
+  // @ts-ignore
   const thumbnailTemplate = document.querySelector('#picture').content;
   const fragmentThumbnail = document.createDocumentFragment();
 
@@ -15,19 +17,17 @@ const renderThumbnails = function (quantity) {
   createPhotoDescriptions(quantity).forEach(({url, likes, comments, description}) => {
     const thumbnailItem = thumbnailTemplate.cloneNode(true);
 
-    const imageSrc = thumbnailItem.querySelector('.picture__img');
-    imageSrc.src = url;
+    thumbnailItem.querySelector('.picture__img').src = url;
 
-    const imageLikes = thumbnailItem.querySelector('.picture__likes');
-    imageLikes.textContent = likes;
+    thumbnailItem.querySelector('.picture__likes').textContent = likes;
 
-    const pictureCommentsQuantity = thumbnailItem.querySelector('.picture__comments');
-    pictureCommentsQuantity.textContent = comments.length;
+    thumbnailItem.querySelector('.picture__comments').textContent = comments.length;
 
+    // Находим и слушаем клик по ссылке
     const link = thumbnailItem.querySelector('.picture');
-
     link.addEventListener('click', () => {
-      drawBigPicture(url, likes, comments, description)
+      drawBigPicture(url, likes, comments, description);
+      modalOpen();
     });
 
     fragmentThumbnail.appendChild(thumbnailItem);
